requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.10.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.10.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
